+++ NEWS.md
@@ -1,5 +1,9 @@
 # gdalio dev
 
+* Included `gdalio_matrix()`, `gdalio_array()`, and `gdalio_graphics()` as exported functions. 
+
+* New function `gdalio_format_source()` prints the required code to run to load functions for stars, terra, raster, spatstat objects. 
+
 * Added support for raster, terra, stars objects to `gdalio_set_default_grid()`. 
 
 * New function `gdalio_local_grid()` to easily generate a local grid around a longitude,latitude.